refactor: Dao의 인터페이스를 제거하고 mock 기반 테스트로 변경한다
